Parser class: move parsing of TaskType to TaskType class
The parseNewTask method is considerably long.

Long methods are harder to read. Bugs are also harder to find in
long methods.

Let's move the parsing of TaskType from a String object into a new
method in TaskType class.

This would reduce the length of the parseNewTask method,
making it easier to read and understand. It would also allow
other future classes to use the method as well.