Amended for greater readability.
Such amendments are necessary so that other programmers who come across my forked repository will be able to understand the program.

Removed duplicated code, adding default branches for if,else statements, removed the Duke class which was redundant, simplified several expressions.

Makes the code easier to understand and read, especially for programmers who are not aware of my program.