Add a new remind command
A new feature that reminds the user about the upcoming deadlines and
events, when the user types 'remind' in the command line of the
Bubbles bot. This feature improves on the functionality of the
task tracker bot and thus is able to help users track tasks
in a more efficient manner.

Let's generate a list of upcoming deadlines and events (happening
within a week from the time of command) and list them out to the
user.

Additional information that could be included in the remind message
includes the overdue deadlines and events, that might be included
in the next commit.