Add ability to sum expenses within a Task
Expenses within a Task cannot be summed up.

Calculating expenses without a summing function is very troublesome.

Let's implement the ability to detect and execute a /sum function.
It will find the chosen task, sum all expenses involved with that
Task and return the sum to the user.

Improves user experience as they do not need to sum up expenses
individually.