Add ability to sum all expenses from all Tasks
There is no function to calculate the total expenses in the
TaskList and the user has to personally do it.

Summing all adds convenience to the user and is a typically
sought after function when expenses are involved.

Let's implement a function that makes use of the pre-existing
sumExpense function in each Task to sum total expenses in the
whole TaskList.

No Task and expense will be missed by looping through the whole
list and it adds a layer of convenience to the user.