Refactor code to improve code quality
The filtering of user input command keywords and the subsequent actions
taken to create a command is being done through one entire if-else
block.

This is difficult to expand upon in a manageable manner in the future
when there might be more commands available.

Filtering of user input keywords is being achieved through an if-else
block, but the subsequent actions taken after the filtering process
is being extracted into separate commands.

This will allow future command preparations to be easily added into
the parser in the form of a separate method.