No assertions statements.
There is no code in place to indicate possible bugs.

Assertion statements have been added to Duke class to ensure user input
and program output are not null.

The impact of assertions on performance is low and provides a safety
net to ensure the user input and program output are as expected.