Add assertions to the code
Assertions are not used in the code.

Assertions are helpful in determining whether assumptions hold true and if mistakes have been made in the code.

Let's add some assertions to the deleteItem and getTask method in TaskList class. We will assert that the indices provided to these methods are within bounds.

The ability to detect if the index of the item to be retrieved/deleted is out of bounds is important as it suggests that the arguments passed to the methods are incorrect.