Add assertions
FindCommand class does not check for cases where the task list containing keywords might be longer than the original task list.

A keyword task list that is longer than the original task list suggests a flaw in the logic or even possibly an infinitely long list is being created.

An assert expression is added before returning the keyword task list.

This helps to ensure that, in the event of an infinite loop, the program stops if keyword task list begins to grow beyond the original task list.