Improve code quality
The current code does not adhere to the coding quality standards.

The Single Layer of Abstraction (SLAP) rule is violated and there is nested code. Moreover, some class names are inappropriate considering their purpose.

Let's rename the UserInterface class as ResponseGenerator, as the purpose of the class is to generate String responses based on the user's input. In addition, let's reorganize the code in the Parser class to reduce nesting.

Reduced nesting of code and appropriate class naming will allow readers to better understand the code, and also allow us to better maintain the code.