Add support for tagging tasks
Tasks does not support tagging.

Tagging tasks allows grouping of tasks with tags, and searching for a tag will return all tasks with that tag.

Let's,
Add a field in the Task class to contain the tag.
Add support for tagging the tasks.
Add Support for searching of tasks by a tag.