Addition of the scheduling function
Addition of the TimeComparator class and the Schedule class. Additional
method to the taskList

Provide the users with a compiled list of events and deadline for the
day they have inputted

Created the schedule class to manage the processing of the data. The
Timecomparator class is used to sort the ArrayList of tasks.

Ensures greater abstraction from the TaskList class and the DukeMan
class.