Add ability to specify command shortcuts
Users can only use the commands specified in the program.

Adding the ability for users to specify their own shortcuts
(or aliases) for commands will improve the user's workflow by
making the syntax more friendly overall.

Let's
* Add a class to handle aliases
* Add methods to save and load aliases to and from a config file
* Modify certain methods in TaskList to take into consideration
aliases of variable length

Having a dedicated class to handle aliases will ensure that
other classes would not get too long, as well as serving as
an encapsulation of alias-related functionality so that any
future modifications made to how aliases are handled will
not cause a cascade of modifications in other classes.