Storage: change the default case for load to return empty list instead.
The expected behaviour for load is to return an empty list should the
formatting have any errors, as opposed to return a list with some tasks.

Changing the default case to return an empty list handles the case as
mentioned above.

Let's change the default case for load to return an empty list.