Add Java assertions
Duke does not utilise assertions to verify program state during runtime.

Assertions should be used to allow for checking for bugs.

Let's add assertions into the code where appropriate, such as when
reading from storage file, and when processing commands.

The use of assertions will allow checking for possible bugs in the code.