Add assertions
Add assertions to check assumptions in code.

Assertions are added in the Storage class in order to check that the data we load is in a correct/good format.

Assertions are also added in Duke to check for index out of bounds.

These are put at specific points so that if they are caught, we know that something has gone wrong with the parsing of data, either in loading from disk or there are logical errors in the data loaded.