TaskList: Replace numOfTasks with tasks.size()
numOfTasks tracks number of items in the TaskList, which can be
trivially derived from tasks. This results in needing to maintain
an unnecessary counter which is a source for bugs.

Let's replace all instances where number of tasks is needed from using
numOfTasks to directly querying the number of tasks based on the size of
TaskList.

This results in no change to functionality while removing the space for
bugs outlined earlier.