Change DialogBox format
The current printout on the gui makes the user's input look similar
to Duke's output. This makes it seem like the user is chatting with
another person, but this is not the case. There is also no clear
demarcation of where the user's text is.

Changing the format of the output to include a clear textbox and
having Duke's response look different from the user's input will
make it clear that the user is chatting with a bot.

Let's add textboxes and change the color of the textboxes to
differentiate between the user's input and Duke's responses.