Handle malformed user input in Parser.
When user inputs a task ID to delete, if the input ID is not associated
with a task then Duke would throw an error where the user's input would
be ignored while the stacktrace would be printed to console.

Let's handle for such cases of malformed ID input by checking that the
ID is within the range of ID values in our list before trying to delete
the task requested.

This way we would prevent errors from being thrown by attempting to
access an item that does not exist in our tasks arraylist.