Add equals method to TaskList
TaskList did not have an equals method, preventing other classes from
checking if their TaskList is equal to another TaskList.

Let's add an equals method to TaskList that compares that the items
within TaskList are identical to that of another TaskList.

This way Storage would be able to compare if serialised TaskList is
identical to the original TaskList input