Add Help function
Duke has no help function to provide in-App guidance to users should
they be unfamiliar with the application. This is an issue as there will
be higher barriers to entry for new users.

Let's add a help function that lets users see all the commands they can
use within the application. This will significantly reduce the barriers
of entry for new users while simulataneously giving more experienced
users the opportunity to remind themselves of the commands and format of
the commands that they can use.