Fix Parser and improve code quality
Parser currently is buggy with the new introduction of friendlier syntax.

Code needs to be modified so user input will not cause any bugs and will
be accurate.

Abstracted methods out of main parse method and changed methods that used
substring to split.

This improves code readability as well as cause it to be less buggy as user
input is not used separated by a fixed number but by splitting by words.

This commit helps improve SLAP in this program as well as arrow headed style of coding.