duke package : Add assert statements
There is no assertion statements to check if the code behaves
according to my assumptions.

Assert statements help to validate assumptions made about the
program and flag out any assumptions that do not hold.

Let's add assert statements to check if variables are correctly
assigned before they are operated on and the variables take on
reasonable values.