Add assertions that check that the indexes of tasks to be deleted or marked as complete are at least 1.
Should the value of the indexes be less than 1 at the point the
assertions are made be less than 1, the program would crash subsequently
because we would be trying to access a negative index of an ArrayList.
As a first line of defense, we have tried to use exceptions to warn
users should they enter inappropriate index values. The assertions are
put in place in case we (as programmers) fail to use exception handling
in certain cases.

Let's add assertions that terminate the program (and display messages
explaining the terminations) should the indexes be less than 1.

Using assertions is a good choice because their impact on performance
is low and worth the additional safety provided.