AddCommand class: newTask is initiated with null
There is no null check for newTask before moving on to add the tasl
into taskList.

An assertion will detect cases where no task happens to be initialised
and replace the null value.

This will verify the correctness of the program flow when Duke runs
and show the error message.