Extract out checkNumOfArguments() and checkIndexArgument() from the variations of checkValidityCommand() methods.
Certain behaviors are shared among several variations of
checkValidityCommand() methods, which violates the DRY principle.

Let's extract out the common behaviors into separate methods. For
example, all the checkValidityCommand() variations check for an
appropriate number of arguments in the input. Therefore, we abstracted
out a checkNumOfArguments() method and use it in the
checkValidityCommand() methods.

Extract method is a form of refactoring that enhances the readability
and extensibility of the code.

Refer to
https://nus-cs2103-ay2122s1.github.io/website/book/refactoring/what/
for an explanation of extract method.