Add DoWithinPeriodTask
Add a new Task named DoWithinPeriodTask

A DoWithinPeriodTask indicates a task that contains a period, or specifically a start and end date to indicate the time period during which a task can/should be completed.

A new subclass of Task has thus be added to implement this, as this is a perfectly valid Task.

This new class can be saved to disk, formatted and marked as completed as per any other Task.