Add C-DetectDuplicates
When adding a task to the task list, the task list will check whether
the task currently exists. If so, it will prompt the user and the task
cannot be added.

To determine if the task exists, Task objects are compared with the
equals() method. Equality of Task objects is defined as:
*ToDo: String objects for the ToDo name are equal
*Event: String and LocalDate objects are equal
*Deadline: String and LocalDate objects are equal