Improve CodeQuality of the project
The executeCommand method has some common behaviors for certain command types
(e.g. getting the index) as well as can be refactored
(e.g. adding task messages).

The common behaviors cause code duplicaiton and the method is long and should
be refactored.

Extracting the common behavior into a function allows use to use polymorphism
when dealing with the various kinds of Tasks. For example, displaying add
message for the different kind of Tasks with one function.

Let's pull up the common behavior of displaying the message into a new
function displayAddMessage and also refactor the code for adding new
deadlines and events.