Add assertions to main codebase
A good number of functions and methods in Duke utilize some form of user input which is prone to errors.
This is particularly true for methods that deal with parsing of user data.

Ensuring certain assumptions about the code holds will help with ensuring the correctness of the code now
and in future.

As a step towards such an assurance, let's add assertions to the codebase, in particular the Parser and Tasklist
classes, in particular for methods that utilize user input. Doing so will help ensure the correctness of the code.