Add full help functionality.
Users not able to get help on how to
use the commands or the entire dukeBot itself.

Due to this, there is alot of confusion regarding how to operate
the duke bot

Help Command is added to assist the users and give them a brief overview
commands that the current dukebot supports. Also, when the dukebot first starts,
a reference to the help command appears.