Add more OOP (Parser, Commands)
There are many different kinds of commands Katheryne can parse, which are differentiated
by a command word. The process to parse these commands are similar, with certain differences.

Separating these out into different classes, and putting all of the parsing code into it's own
class, increases the readability and the abstraction of the code. It also follows SLAP
principles.

This commit extracts out each command into it's own command under an abstract command class, and
modifies the main code to use this abstracted form. It also removes unused code blocks and further
groups methods which require communication with the user into the Ui class.

It also makes the commands case insensitive, for the user's convenience.