Improve the GUI and fix the creation flow of the GUI
The GUI is at the minimum level where an input is replied with a
generated response and there are basic features that the user can see.
The creation flow of the GUI is standard and all styled within the code itself.

It is not very visually pleasing and does not give unique responses to inputs.
The creation flow of the GUI can be changed to fit the abstraction of the bot
and styling is hard coded within the Window class.

Changing how the bot responds will help to increase interactivity with
the user. Abstracting some parts of the GUI creation to different classes
will make code more understandable and using .fxml files to style the
GUI will be simpler as well.