Add learn and unlearn commands for the user to teach Alice or make Alice forget about certain custom commands.
Alice would now be able to respond to user with specific phrase once the user have taught Alice what to respond.
This could prove to be useful for user to teach Alice a specific phrase and to respond accordingly.

The update also comes with unlearn functionality to make Alice more intuitively usable.

Note:
- User cannot teach Alice the command that Alice already know as Alice will respond back that the vocabulary user is trying to teach is untouchable
- User cannot teach the vocabulary that Alice already knows
- User cannot unteach/make Alice unlearn a vocabulary that she does not recognize or is a default command