Fix codebase's code quality
There are several methods in Duke that do not obey best practices
for coding quality, with several lengthy methods, particularly in
Parser.

Refactoring the relevant portions makes for more readable and
better documented code, that improves the ease of future
development.

To this end, let's refactor the code base by abstracting repeated
portions of code into separate methods, and shortening the lengthy
"parse" method using switch statements and enums. This will help
improve the quality of the code overall.