Change implementation of Undo
The undo implementation for add / delete was previously different from the done, where save states are used in the former
and manually the completion of tasks in the latter.

Added a method to create deepcopy of arraylist of task, and set a save state for all 3 commands instead.

Standardises the method of saving states throughout the app.