Add JavaDocs to the FindCommand and make edits to other Command classes
There are no JavaDocs for FindCommand and other Command classes are not
updated to reflect the return value changes.

JavaDocs need to be rectified to properly reflect the new updates due to
JavaFX.

Let's edit the JavaDocs to reflect the changes to the Duke program since
the last time JavaDocs were edited.