Improve the GUI
The GUI uses the same colors for the user and duke. In addition, the pictures used look unprofessional.

Improving the GUI to make it more polished helps users to differentiate between user and duke messages and become more satisfied by the aesthetics.

Let's improve the GUI with better colors and images.