Add assertions to document important assumptions in the code
None of the existing classes have assertions to detect the source of
logical flaws in the code.

The addition of assertions will increase the probability of identifying
bugs in the program that are caused by violation of certain assumptions
in the program logic.

Let's add assertions to the relevant parts of the code to define
certain assumptions about the program state.

Using assertions is greatly advantageous as assertions are easy to
implement and help document and verify key assumptions that should
hold at various points in the code.