Improve code quality for better readability and reusability.
Large chunks of code in methods.
Some String literals are duplicated in the code.

Large chunks of code hinder readability and it is better not to have
duplicated code.

Refactoring the methods with large number of lines into multiple shorter
methods will help to improve readability and will improve ease of
changing the code in the future.