Implement Update extension feature
Users are not able to update tasks after they have added.

This becomes quite a hassle as users must delete and add tasks again
to perform an update.

Implementing a feature to allow easier updates of tasks can improve
ease of use for the users. Users will be able to update either the
task's description, date & time or the entire task.