Add support for DoAfter tasks
Support already exists for a variety of task types, including
ToDos, Deadlines and Events, with the ability to add datetime
information for relevant task types.

Adding support for DoAfter tasks provides end-users with more
flexibility in task management when using the app.

To this end, let's implement support for a "doafter" command which
allows users to create a DoAfter task by
* Updating the Parser to parse "doafter" commands
* Updating PersistentStorage to read/write DoAfter tasks to file
* Adding a new DoAfterCommand type