Add PictureEditor class, and remove unnecessary code
Code has potential to have another class, PictureEditor class abstracted out, that
deals with the editing of the pictures to be shown. Additional, there is some code
that is redundant and that can be removed.

Extract PictureEditor class, and clean up code.

This is done to improve OOP and increase readability of the code overall.