Implement A-BetterGui
To appeal to the users of this program, the GUI has been designed to be
more attractive and at the same time, fixed certain bugs that did not
allow the program to work as desired.

Things that has been done
*GUI initialises the window with a welcome message
*GUI will not truncate the text if it exceeds the preset height
*Amended the format to be printed on GUI and on console

It is done this way to ensure that the program behaves ideally.