Refactor code
Duke stores the string to be rendered when HelpCommand is executed in a
seprate file. However, when deployed to jar file, this txt file cannot
be read.

Let's create a separate class called HelpProvider so that the String
can still be retrieved when the jar file is used in another directory.