Add functionality for Duke to return statistics.
Duke can add and delete tasks and mark tasks as done, but sometimes the list of tasks get very long and quite unwieldy to handle.

Providing the ability for the user to view statistics allows the user to quickly see their progress at a glance.

To provide such an ability, three new commands have been added with the purpose of showing the user such statistics.