Improve GUI
The graphical user inteface looks simpler and better.

The GUI of the original version looks like something from the '90s.
Users might refuse to use it because of the bad GUI design.

Modifying the design of dialog boxes, the size of the whole window,
and the font of 'Send' button.

Changing the parts can make the whole GUI look simple and modern.

Refer to seniors' work
https://nus-cs2103-ay2021s1.github.io/website/admin/ip-showcase.html