Add assertions to Duke application.
Currently, all classes in the application do not have assertions that
verify the assumptions made about the program state. As a result, there
are many places that are prone to NullPointerExceptions.

As a result, assertions are added to the actions, Request, Response and
Duke classes so that the runtime can verify these assumptions and flag
them out when breached.

Classes affected:
* actions
  * AddTask
  * CompleteTask
  * DeleteTask
  * FindTasks
  * ListTasks
* Request
* Response
* Duke