Feat: Add the find command
When the user enters many tasks, it may be difficult to find the one you are looking for.

Using the find command allows users to find tasks more easily.

In the future, we might look at sorting tasks when they are returned, so that it is organised
more neatly and tasks can be even easier to find.