Add assertions feature
The assertions feature addresses important assumptions that should
hold in the code.

A program that terminates the program early once a bug is observed
is safer to use as it prevents the error from escalating further.

Let's include the assertions feature to define assumptions about the
program state so that the runtime can verify them.