Improve GUI of Duke
The GUI of Duke needs to add value to the program, with one such way
being added readability.

Adding visual elements to the GUI that help with the readability would
help with the overall user experience.

To this end, let's demarcate the different dialog boxes used (by Duke vs
the user) through the use of different colours, and clip the profile
pictures into a circle for aesthetics.