Add Result class to project
Currently, the execute method for all commands returns a TaskList
object.

This is a sign of inadequate abstraction, as it does not make much
logical sense for all commands to return a task list, especially for
alias-related commands that do not modify the task list. In addition,
AliasHandler and TaskList currently contain a message attribute to
store the output message, which is a potential sign of repeated code.

Let's,
* create a new class, Result, that stores an AliasHandler object and
TaskList object, as well as an output message string.
* make all commands return Result instead of TaskList.
* remove the message attribute from TaskList and AliasHandler.

This would provide another layer of abstraction, as well as making
the code easier to understand. (The return value of executing a
Command is a Result.) It also reduces code duplication, as
there is now only one message attribute. (In Result)