Implement bug fix for loading tasks from file
Task strings stored in the persistent file were not validated before
loading into the Duke application, which led to the application
crashing if a malformed persistent storage file was provided.

Validating the task strings stored in the persistent file improves the
resilience of the application.

To this end, let's
* implement a validation function in PersistentStorage
that validates task strings
* implement checking in the Duke GUI to inform the user should the
persistence file not be successfully loaded