Tidy up implementation and work on GUI
GUI is very basic and not customised to the implementation of Duke
envisioned. Some methods are also too long and should be separated out
into other smaller, more self-contained methods.

Let's put in new background and user and duke images into the GUI, as
well as set backgrounds for the dialog boxes for the text to contrast
more against the overall background.

Let's also split the method in the Parser class into smaller methods so
that it complies with the coding conventions and is easier for readers
to understand.