Add assertions for key steps of programme.
There are currently no checks for correctness of the program.

Assertions need to be  setup to ensure the correctness of the code at keypoints in the chatbot,
this includes the the Chatbot package as well as the implementation of the Chatbot class.

The following has been done to achieve this:
* check the chatting status, ensure user input handlers are not triggered when chatbot is inactive.
* check the output stream was modified after utilizing the guiInterpret, failure of switch back
is potentially application breaking
* check the loading of tasks is done at the beginning, ensure task lists start from a clean slate.

Java's assert is used so that application can ignore such assertions after deployed.