Insert JavaDoc comments and other GUI changes
JavaDoc comments are out of date and some methods do not have JavaDoc
comments. In addition, the text for the GUI is a bit plain.

Let's add JavaDoc comments to the methods that don't have them and
update the other JavaDoc comments so that the code is understandable to
readers.

Let's also update the text and font of the GUI to make it a bit more
customised to the Duke program.