Add a 2 second delay after the bye command
Duke does not have a delay after the bye command is executed, which
means that the program closes immediately after the user sends the
command. It would feel smoother if there was a slight delay.

Let's add a 2 second delay after the bye command is sent to achieve the
purpose stated above.