Add option to archive current list of tasks
There is no way to archive tasks.

This causes the list of tasks to continue to grow over time.
It is not user friendly to view a long list of tasks
in the GUI.

Let's provide a command for the user to specify the filename
to which the current tasks will be archived to.