Update EventTask to check for time clashes
B-DetectAnomalies

EventTasks currently only hold dates as data. Let's add time as well. EventTask now accepts datetime in the format yyyy-MM-dd HH:mm.

Users may want to know if they are adding an event that clashes with their other events. Let's add functionality to inform them if there is an overlap in the timing of the event they are adding.