TaskList class: Apply SLAP to shorten methods of adding tasks
The addTask method implements the addition of
multiple types of tasks (i.e. ToDo, Deadline, Event).

This violates the Single Level of Abstraction principle and
makes the code difficult to understand or debug.

Implementing the add task methods separately allows us to call different
methods when adding a task of a particular type (e.g. addTodo).

Let's implement addTodo, addDeadline, and addEvent as separate methods.

This reduces the risk of calling the wrong add task method.