Add assertion checks for TaskList
Currently the Task List does not check the index for getTask,
completeTask and deleteTask functions.

Currently the Storage class does not have a check for the TastList being
uninitialised in the Load function

Let's add assertion checks verify if the index is within the limit
(0 - Size) in the Task List class and also place a null check for the
Task List in the Storage class for the Storage.load function.