Add in a check for data file
Data file under the default path of ./data/data.txt
may or may not exist.
There is no need to create it if it already exists.

The Storage class always creates the data file.
If the file already exists, it will enter the catch block and
print out error message when there is no error at all.

Add in a check before creating the file. Only create it when
it is not yet existent.