Implement and use toString() of TaskList
Some classes currently print of tasks by first obtaining the list of individual Tasks from TaskList.

This is bad, because a function that is done on TaskList should be performed by TaskList itself.

Using the SLAP principle, lets move getting the string of tasks into TaskList itself, and printing that to the UI directly instead.