Incorporate Katheryne logic
This commmit adds in the logic which was present in the CLI but not the GUI,
such as parsing the commands and printing out visually responses. It needs to
change for the GUI to be an appropriate substitute.

I have added the getResponse method to all commands, which returns a string
which will be handled by the GUI. I have also used Platform.exit() to close
the GUI during the bye command in handle user input.

I have updated some of the javadocs which are now deprecated as the methods
pertained to the CLI, which will no longer be actively edited.

There is still a lot to be improved, such as the look of the GUI as well as
the way that the application is closed. I also want to add a greeting message
when Katheryne is opened.