Change find command to show task index
When the find command returns a list of results, it returns them with the items numbered incrementally from 1. This is not helpful for users who want to know what the exact index of the item is on their task list.
Let's change it such that instead of renumbering the search results, it shows their exact index number from the task list.