Add Assertions feature: Assert input given by user will always be valid
Use Assertions feature (not JUnit assertions) to document important assumptions that should hold at various points in the code.

Add two assert lines, one is in Parser, to ensure that Parser will always parse userInput correctly, and another is on Command, to make sure
the message input can be executed. We do so because it can satisfy our need for safe work in Duke.