Add Assert functions
Added 2 assert functions to the program. 1 was added to the Duke class
to ensure that when the function goes through all of the if-else
statements, it is because no valid input has been inserted by the user.

The second assert function was added to the Storage class. It is used to
detect if there were any invalid task type that somehow managed to get
into the saved text file.