Implement A-CodeQuality of individual project
Currently the code for the bot across multiple classes is not entirely refined and is a bit messy.

The code is improved based on the code quality topics learnt in CS2103T so as to make it more readable and maintainable.

This includes separating parts of long methods out, adding more javadoc comments, and minor bug fixes and optimisations.