Implement C-Sort extension for individual project
Right now, there is no way to manipulate the task list
besides adding and deleting a task.

This can make the task list messy and difficult to track tasks
after multiple insertions or deletions.

A sort function is implemented which allows the user
to sort the task list by task and date/time.

This allows them to separate out different types of tasks,
and see which deadlines/events are coming up soon.