Add Log capability to Storage Class
To reduce clutter on chat interface and to avoid
distracting the user with error messages, these messages
should be placed elsewhere, where they are noticeable.

Add Log capabilities to the Storage class.

Any error message with regards to File operations and Storage
class would be logged to the console and stored in a file
'storageLog.log' as well.