Clean up codebase and neaten code
The Parser in Duke was too long and led to unsightly code. In
addition, validation was done by a separate function that led to
unnecessary code.

To fix this, more abstractions in the parsing of the commands in
the Parser class could lead to cleaner code.

To this end, let us
* abstract each switch case in the Parser class into its own
function

A bug where resizing the Duke GUI would lead to unexpected
window behaviours was also fixed. The GUI is now non-resizable.